requirements.txt
streamlit==1.28.1
pandas==2.1.3
plotly==5.18.0
psycopg2-binary==2.9.9
